fix vs warning

+++ saber/cbd.c
@@ -29,7 +29,7 @@ void PQCLEAN_NAMESPACE_cbd(uint16_t *r, const unsigned char *buf) {
     int i, j;
 
     for (i = 0; i < SABER_N / 4; i++) {
-        t = load_littleendian(buf + 3 * i, 3);
+        t = (uint32_t) load_littleendian(buf + 3 * i, 3);
         d = 0;
         for (j = 0; j < 3; j++) {
             d += (t >> j) & 0x249249;
@@ -80,7 +80,7 @@ void PQCLEAN_NAMESPACE_cbd(uint16_t *r, const unsigned char *buf) {
     int i, j;
 
     for (i = 0; i < SABER_N / 4; i++) {
-        t = load_littleendian(buf + 5 * i, 5);
+        t = (uint32_t) load_littleendian(buf + 5 * i, 5);
         d = 0;
         for (j = 0; j < 5; j++) {
             d += (t >> j) & 0x0842108421UL;
